Moved ActivityExecutionContext ot activity package
